Created the player testing branch
I've tagged the collectible and the Obstacle, I've set up the movement for the player and I've laid out the framework for the collision detection I'm having trouble implementing it, but I just have to get the detection to happen and the rest is easy. The player can move left and right and can jump, the player can also move left and right in the air as well.